✨ feat(MockingTest.php): add unit test for mocking ClassToBeMocked
The MockingTest.php file was added to the phpunit/complex/tests directory. This file contains a unit test for mocking the ClassToBeMocked class. The test creates a mock object of ClassToBeMocked using the createMock method provided by PHPUnit. It then sets an expectation that the methodToBeMockedA should return 'C' when called on the mock object. The test asserts that the methodToBeMockedA returns 'C' and that the methodToBeMockedB returns null when called on the mock object.